refactor(app-shell): 拆分工作区流转编排
